adds check in startup script and sudo-deps to check if the appliance is online

+++ packer/scripts/02-deps.sh
@@ -24,10 +24,12 @@ CRUCIBLE_VARS=$(cat <<EOF
 #!/bin/bash 
 export APPLIANCE_VERSION=
 export APPLIANCE_IP=
-APPLIANCE_ENVIRONMENT=
+export APPLIANCE_ENVIRONMENT=
+export IS_ONLINE=
 EOF
 )
-
+IS_ONLINE=$(curl -s --max-time 5 ifconfig.me >/dev/null && echo true || echo false)
+echo "IS_ONLINE: $IS_ONLINE"
 # Detect Mac and use greadlink
 readlink_cmd="readlink -m"
 if [[ "$OSTYPE" == "darwin"* ]]; then
@@ -70,20 +72,22 @@ if [[ ! -f /etc/profile.d/crucible-env.sh ]]; then
     sudo sed -i "/APPLIANCE_VERSION=/c\export APPLIANCE_VERSION=\\$APPLIANCE_VERSION" /etc/profile.d/crucible-env.sh
     sudo sed -i "/APPLIANCE_IP=/c\export APPLIANCE_IP=\\$APPLIANCE_IP" /etc/profile.d/crucible-env.sh
     sudo sed -i "/APPLIANCE_ENVIRONMENT=APPLIANCE/c\export APPLIANCE_ENVIRONMENT=APPLIANCE" /etc/profile.d/crucible-env.sh
+    sudo sed -i "/IS_ONLINE=/c\export IS_ONLINE=\\$IS_ONLINE" /etc/profile.d/crucible-env.sh
 else
     echo "Setting APPLIANCE_VERSION to $APPLIANCE_VERSION in /etc/profile.d/crucible-env.sh"
     sudo sed -i "/APPLIANCE_VERSION=/c\export APPLIANCE_VERSION=\\$APPLIANCE_VERSION" /etc/profile.d/crucible-env.sh
     sudo sed -i "/APPLIANCE_IP=/c\export APPLIANCE_IP=\\$APPLIANCE_IP" /etc/profile.d/crucible-env.sh
     sudo sed -i "/APPLIANCE_ENVIRONMENT=/c\export APPLIANCE_ENVIRONMENT=APPLIANCE" /etc/profile.d/crucible-env.sh
+    sudo sed -i "/IS_ONLINE=/c\export IS_ONLINE=\\$IS_ONLINE" /etc/profile.d/crucible-env.sh
 fi
-
 
 ######################
 ###### Update OS #####
 ######################
-sudo apt-get update -y && sudo NONINTERACTIVE=1 apt-get dist-upgrade --yes && sudo apt-get autoremove -y
-sudo NONINTERACTIVE=1 apt-get install -y build-essential avahi-daemon jq nfs-common sshpass postgresql-client make logrotate git unzip apache2-utils
-
+if $IS_ONLINE; then
+    sudo apt-get update -y && sudo NONINTERACTIVE=1 apt-get dist-upgrade --yes && sudo apt-get autoremove -y
+    sudo NONINTERACTIVE=1 apt-get install -y build-essential avahi-daemon jq nfs-common sshpass postgresql-client make logrotate git unzip apache2-utils
+fi
 ########################
 ##### Configure OS #####
 ########################
@@ -161,6 +165,7 @@ CURRENT_IP=$(ip route get 1 | awk '{print $(NF-2);exit}')
 APPLIANCE_VERSION=${APPLIANCE_VERSION:-$(cat /etc/appliance_version)}
 DOMAIN=${DOMAIN:-crucible.local}
 
+
 # Delete old entry
 sudo sed -i "/$DOMAIN/d" /etc/hosts
 msg="Entry being added in hosts file. entry: '$CURRENT_IP    $DOMAIN'"
@@ -173,57 +178,56 @@ msg="Entry update in host file: /etc/hosts '$CURRENT_IP   $DOMAIN'"
 ################################
 ##### Install Dependencies #####
 ################################
+DIST_DIR=/home/crucible/crucible-appliance/dist
+VAULT_VERSION="1.15.2"
+ARGOCD_VERSION="2.13.1"
 
+if [ $IS_ONLINE ]; then
+    echo "Downloading Binaries"
+    curl -z $DIST_DIR/generic/k3s -Lo $DIST_DIR/generic/k3s "https://github.com/k3s-io/k3s/releases/download/v1.29.1%2Bk3s1/k3s"
+    curl -z $DIST_DIR/generic/k3s-install.sh -Lo $DIST_DIR/generic/k3s-install.sh "https://get.k3s.io/" && chmod +x "$DIST_DIR/generic/k3s-install.sh"
+    curl -z $DIST_DIR/generic/yq -Lo $DIST_DIR/generic/yq "https://github.com/mikefarah/yq/releases/latest/download/yq_linux_amd64"
+    curl -z $DIST_DIR/generic/kubectl -Lo  "https://dl.k8s.io/release/v1.31.3/bin/linux/amd64/kubectl"
+    curl -z $DIST_DIR/generic/k9s_linux_amd64.tar.gz -Lo $DIST_DIR/generic/k9s_linux_amd64.tar.gz "https://github.com/derailed/k9s/releases/download/v0.32.7/k9s_Linux_amd64.tar.gz"
+    curl -z $DIST_DIR/generic/argocd-linux-amd64 -Lo $DIST_DIR/generic/argocd-linux-amd64 "https://github.com/argoproj/argo-cd/releases/download/v$ARGOCD_VERSION/argocd-linux-amd64"
+    curl -z $DIST_DIR/generic/vault_${VAULT_VERSION}_linux_amd64.zip -Lo $DIST_DIR/generic/vault_${VAULT_VERSION}_linux_amd64.zip "https://releases.hashicorp.com/vault/${VAULT_VERSION}/vault_${VAULT_VERSION}_linux_amd64.zip"
+    curl -z $DIST_DIR/generic/go1.22.5.linux-amd64.tar.gz -Lo $DIST_DIR/generic/go1.22.5.linux-amd64.tar.gz "https://dl.google.com/go/go1.22.5.linux-amd64.tar.gz"
+    curl -z $DIST_DIR/generic/helm-v3.16.3-linux-amd64.tar.gz -Lo $DIST_DIR/generic/helm-v3.16.3-linux-amd64.tar.gz https://get.helm.sh/helm-v3.16.3-linux-amd64.tar.gz
+fi  
 # Install yq
-sudo wget https://github.com/mikefarah/yq/releases/latest/download/yq_linux_amd64 -O /usr/bin/yq && sudo chmod +x /usr/bin/yq
+sudo mv $DIST_DIR/generic/yq /usr/bin/yq && sudo chmod +x /usr/bin/yq
 
 # Install K3s
 sudo mkdir -p /etc/rancher/k3s
 mkdir -p ~/.kube
-# sudo echo "nameserver 10.0.1.1" >> /etc/rancher/k3s/resolv.conf
 sudo echo "$MIRRORS" > /etc/rancher/k3s/registries.yaml
-curl -sfL https://get.k3s.io | INSTALL_K3S_VERSION="v1.29.1+k3s1" K3S_KUBECONFIG_MODE="644" INSTALL_K3S_EXEC="server --disable traefik --embedded-registry --etcd-expose-metrics --cluster-init --prefer-bundled-bin --tls-san ${DOMAIN:-crucible.local}" sh -
+sudo mv $DIST_DIR/generic/k3s /usr/local/bin/k3s && sudo chmod +x /usr/local/bin/k3s
+INSTALL_K3S_SKIP_DOWNLOAD=true INSTALL_K3S_VERSION="v1.29.1+k3s1" K3S_KUBECONFIG_MODE="644" INSTALL_K3S_EXEC="server --disable traefik --embedded-registry --etcd-expose-metrics --cluster-init --prefer-bundled-bin --tls-san ${DOMAIN:-crucible.local}" $DIST_DIR/generic/k3s-install.sh
 mkdir ~/.kube
 sudo cp /etc/rancher/k3s/k3s.yaml ~/.kube/config
-sed -i 's/default/crucible-appliance/g' ~/.kube/config
-sed -i "s/127.0.0.1/crucible.local/g" ~/.kube/config
+sed -i "s/default/\${DOMAIN}/g" ~/.kube/config
+sed -i "s/127.0.0.1/\${DOMAIN}/g" ~/.kube/config
 sudo chown -R $SSH_USERNAME:$SSH_USERNAME ~/.kube
 chmod go-r ~/.kube/config
 
 # Install Kubectl
-curl -LO "https://dl.k8s.io/release/$(curl -L -s https://dl.k8s.io/release/stable.txt)/bin/linux/amd64/kubectl"
-install -o root -g root -m 0755 kubectl /usr/local/bin/kubectl
-rm kubectl
+install -o root -g root -m 0755 $DIST_DIR/generic/kubectl /usr/local/bin/kubectl
 
 # Install Helm
-curl -fsSL -o get_helm.sh https://raw.githubusercontent.com/helm/helm/main/scripts/get-helm-3
-chmod 700 get_helm.sh
-./get_helm.sh
-rm -rf ./get_helm.sh
+tar -C /usr/local/bin -xzf "$DIST_DIR/generic/helm-v3.16.3-linux-amd64.tar.gz" linux-amd64/helm --strip-components=1 
+sudo chown root:root /usr/local/bin/helm
 
 # Install K9s
-curl -sLo k9s_linux_amd64.deb "https://github.com/derailed/k9s/releases/download/v0.32.5/k9s_linux_amd64.deb"
-sudo NEEDRESTART_MODE=a apt-get install --yes ./k9s_linux_amd64.deb
-#tar -xvzf k9s.tar.gz
-#sudo -s mv ./k9s /usr/local/bin/k9s
-rm k9s_linux_amd64.deb
+sudo tar -C /usr/local/bin -xzf $DIST_DIR/generic/k9s_linux_amd64.tar.gz 
 
 # Install argocd-cli
-VERSION=$(curl -L -s https://raw.githubusercontent.com/argoproj/argo-cd/stable/VERSION) 
-curl -sSL -o argocd-linux-amd64 https://github.com/argoproj/argo-cd/releases/download/v$VERSION/argocd-linux-amd64 
-sudo install -m 555 argocd-linux-amd64 /usr/local/bin/argocd 
-rm argocd-linux-amd64
+sudo install -m 555 $DIST_DIR/generic/argocd-linux-amd64 /usr/local/bin/argocd 
 
 # Install go
-mkdir -p dist/tools
-curl https://dl.google.com/go/go1.22.5.linux-amd64.tar.gz -o dist/tools/go1.22.5.linux-amd64.tar.gz
-sudo rm -rf /usr/local/go && sudo tar -C /usr/local -xzf ./dist/tools/go1.22.5.linux-amd64.tar.gz
+sudo rm -rf /usr/local/go && sudo tar -C /usr/local -xzf $DIST_DIR/generic/go1.22.5.linux-amd64.tar.gz
 
 # Install Vault
-VERSION="1.15.2"
-mkdir -p dist/tools
-curl https://releases.hashicorp.com/vault/${VERSION}/vault_${VERSION}_linux_amd64.zip -o dist/tools/vault_${VERSION}_linux_amd64.zip
-unzip dist/tools/vault_${VERSION}_linux_amd64.zip -d /usr/local/bin/
+unzip $DIST_DIR/generic/vault_${VAULT_VERSION}_linux_amd64.zip -d /usr/local/bin/
 vault --version
 
 # Install Brew
